+++ notes
@@ -1,0 +1,17 @@
+TODO:
+  - Webworker vs. Node.js:
+    - Adapter Node.js - Webworker: Task & Channel API (blocking Channel)
+    - Statesharing?
+  - map SFN to webworkers -> start them and post result to next until done
+  - check if updated paths are still necessary
+
+
+Node concurrency:
+  - one thread with asynchronous callbacks -> eventloop
+  - concurrency through C++-APIs (browser wpuld have web-APIs)
+  - Node does not provide conucrrency
+
+dynamisches bauen möglich: Skript für Webworker direkt übergeben -> import im Webworker
+
+
+enable: --experimental-modules
