+++ testcases/hello_world/hello.js
@@ -1,9 +1,20 @@
-exports.calc = function (arg) {
+hello = (function()
+{
+  function calc(arg)
+  {
     console.info(`Input: ${arg}`);
     return (arg * 2);
-}
+  }
 
-exports.world = function(arg) {
-    console.info(`Intermediate: ${arg}`);
-    return (arg * 2);
-}
+  export function world(arg)
+  {
+      console.info(`Intermediate: ${arg}`);
+      return (arg * 2);
+  }
+
+  return
+  {
+    calc: calc,
+    world: world
+  }
+})()
